+++ ricaun.Revit.UI.Tests/Items/RevitCreateItemsTests.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using System;
 
 namespace ricaun.Revit.UI.Tests.Items
 {
@@ -18,6 +19,15 @@ namespace ricaun.Revit.UI.Tests.Items
         public void CreatePushButton()
         {
             ribbonPanel.CreatePushButton<BaseCommand>();
+        }
+        [TestCase(3)]
+        public void CreatePushButton_Repeat(int numberOfCommands)
+        {
+            for (int i = 0; i < numberOfCommands; i++)
+            {
+                var pushButton = ribbonPanel.CreatePushButton<BaseCommand>();
+                Console.WriteLine(pushButton.Name);
+            }
         }
 
         [Test]