+++ ricaun.Revit.UI/RibbonStackExtension.cs
@@ -14,6 +14,7 @@ namespace ricaun.Revit.UI
         /// <param name="ribbonPanel"></param>
         /// <param name="ribbonItems"></param>
         /// <returns></returns>
+        /// <remarks>The <paramref name="ribbonItems"/> is divided into groups of a maximum of 3 in each RibbonRowPanel</remarks>
         public static Autodesk.Windows.RibbonRowPanel[] RowStackedItems(this RibbonPanel ribbonPanel, params RibbonItem[] ribbonItems)
         {
             return ribbonPanel.CreateRowStackedItemsWithMax(ribbonItems);
@@ -25,6 +26,7 @@ namespace ricaun.Revit.UI
         /// <param name="ribbonPanel"></param>
         /// <param name="ribbonItems"></param>
         /// <returns></returns>
+        /// <remarks>Each <paramref name="ribbonItems"/> is update to work with RibbonFlowPanel</remarks>
         public static Autodesk.Windows.RibbonFlowPanel FlowStackedItems(this RibbonPanel ribbonPanel, params RibbonItem[] ribbonItems)
         {
             return ribbonPanel.CreateFlowStackedItems(ribbonItems);